Add compassTarget to town level description
This commit resolves the issue #27 where the compass on the first level
points to the shop instead of dungeon entrance.

In addition, this commit enbles a feature where the compass target can
be set to any location on the map, in the future it might be used for
quests to be able to update the compass.

Fix #27