Split WndJournal.updateList() into level/log tabs
In this commit, the method updateList(), that branched according
to the value of showLevels flags, is split into two code segments,
each residing in its respective tab select method. This makes the
flag unneeded. It looks like a lot of changes, but it's basically
cut and paste with some minor changes.