Remove WndJournal.showLevels flag
This removes the boolean showLevels, and from now on, upon opening
the journal window, the first tab (levels) will be shown. This is
different from the behavior before - the last selected tab was
opened. But since the flag is gone, there is no way of knowing which
tab was last opened.